Add support for defining top-level main functions with nested commands
Previously this was not supported since the macro needs to handle this
case specially. The reason is bit subtle, but basically boils down
differences in how top-level classes and methods are encoded:

- classes are declared in a package
- methods are declared in a synthetically-generated 'package class'
  (presumably, because java bytecode supports only top-level class
  declarations)

This means that if a main method looks up annotated command classes in
`this`, then the lookup paths need to be handled differently depending
on whether or not the main method is top-level. There are also some
differences in instantiating the command class.